+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
 ### Fixed
 - [#124] - Automatic section size for table column ([#122]; thanks [@jpcima], [@trebmuh])
+- [#96] - Clean audio buffer initially (thanks [@OPNA2608])
 
 [#122]: https://github.com/rerrahkr/BambooTracker/issues/122
 [#123]: https://github.com/rerrahkr/BambooTracker/pull/123
 [#125]: https://github.com/rerrahkr/BambooTracker/pull/125
 [#124]: https://github.com/rerrahkr/BambooTracker/pull/124
+[#96]: https://github.com/rerrahkr/BambooTracker/issues/96
 
 ## v0.2.0 (2019-04-30)
 ### Added